ingest/requirements.txt
boto3
fastapi
fastparquet
httpx
numpy
pandas
pyarrow
python-dotenv
redis
requests
uvicorn[standard]
pytest-asyncio
